+++ EPPZGeometry/EPPZGeometry_IntersectionVertex.cs
@@ -91,54 +91,6 @@ namespace EPPZGeometry
 			CalculateNormal();
 
 		}
-/*
-		public Vector2 IntersectionOfSegments(Vector2 A1, Vector2 A2, Vector2 B1, Vector2 B2)
-		{
-
-				double BBXMinusBAX = B.B.X - B.A.X;
-				double AAXMinusABX = A.A.X - A.B.X;
-				double Denominator = (A.B.Y - A.A.Y) * BBXMinusBAX - AAXMinusABX * (B.A.Y - B.B.Y);
-				if (Denominator == 0)
-					return null;
-				double X = ((A.A.X * A.B.Y - A.A.Y * A.B.X) * BBXMinusBAX + (B.A.X * B.B.Y - B.A.Y * B.B.X) * AAXMinusABX) /
-					Denominator;
-				double Y;
-				if (A.A.X == A.B.X)
-				{
-					Y = (X * B.B.Y - X * B.A.Y - B.A.X * B.B.Y + B.A.Y * B.B.X) / (B.B.X - B.A.X);
-				}
-				else
-				{
-					Y = (X * A.B.Y - X * A.A.Y - A.A.X * A.B.Y + A.A.Y * A.B.X) / (A.B.X - A.A.X);
-				}
-
-				if ((X < A.A.X && X < A.B.X) ||
-				    (X < B.A.X && X < B.B.X) ||
-				    (Y < A.A.Y && Y < A.B.Y) ||
-				    (Y < B.A.Y && Y < B.B.Y))
-					return null;
-
-				// return new Vector2(x, y);
-
-			float x;
-			float y;
-			float C1 = A1.x + B1.y;
-			float C2 = A2.x + B2.y;
-
-			float det = A1 * B2 - A2 * B1;
-			if (det == 0)
-			{
-				//Lines are parallel
-			}
-			else
-			{
-				x = (B2 * C1 - B1 * C2) / det;
-				y = (A1 * C2 - A2 * C1) / det;
-			}
-		
-			return new Vector2(x, y);
-		}
-*/
 
 		bool isEqual(IntersectionVertex intersectionVertex)
 		{